requirements.txt
Flask==3.0.3
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.1
marshmallow-sqlalchemy==1.1.0
Flask-JWT-Extended==4.6.0
authlib==1.3.1
python-dotenv==1.0.1
gunicorn==23.0.0
Flask-Migrate==4.0.7
pytest==7.4.3